# api/docs/swagger.yaml
basePath: /
definitions:
  file.Visibility:
    enum:
      - private
      - public
    type: string
    x-enum-varnames:
      - VisibilityPrivate
      - VisibilityPublic
...
  handlers.ErrorResponse:
    properties:
      error:
        example: validation error
        type: string
    type: object
...
  handlers.UploadFileResponse:
    properties:
      file:
        $ref: "#/definitions/presenter.FileResponse"
    type: object
...
  presenter.FileResponse:
    properties:
      bucket:
        type: string
      contentType:
        type: string
      createdAt:
        type: string
      id:
        type: string
      originalName:
        type: string
      provider:
        enum:
          - aws
          - minio
        example: aws
        type: string
      sizeBytes:
        type: integer
      storageKey:
        type: string
      storedName:
        type: string
      updatedAt:
        type: string
      uploadedById:
        type: string
      visibility:
        allOf:
          - $ref: "#/definitions/file.Visibility"
        enum:
          - private
          - public
        example: private
    type: object
  presenter.PaginatedFilesResponse:
    properties:
      items:
        items:
          $ref: "#/definitions/presenter.FileResponse"
        type: array
      meta:
        $ref: "#/definitions/presenter.PaginationMetaResponse"
    type: object
...
  presenter.PaginationMetaResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      search:
        type: string
      sortBy:
        type: string
      sortOrder:
        type: string
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  presenter.SignedDownloadResponse:
    properties:
      file:
        $ref: "#/definitions/presenter.FileResponse"
      url:
        type: string
    type: object
...
info:
  contact: {}
  description:
    API server built with Gin, Gorm, Postgres, JWT auth, refresh tokens,
    and Swaggo.
  title: Boilerplate Go Vue API
  version: "1.0"
paths:
...
  /api/v1/files:
    get:
      description: Returns uploaded files. Admin sees all, users see only their files.
      parameters:
        - description: Page number
          in: query
          name: page
          type: integer
        - description: Page size
          in: query
          name: limit
          type: integer
        - description: Search on name, content type, provider, visibility
          in: query
          name: search
          type: string
        - description: Sort field
          enum:
            - createdAt
            - updatedAt
            - name
            - provider
            - size
          in: query
          name: sortBy
          type: string
        - description: Sort order
          enum:
            - asc
            - desc
          in: query
          name: sortOrder
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.PaginatedFilesResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List files
      tags:
        - files
  /api/v1/files/{id}:
    delete:
      description: Deletes the object from storage and removes its metadata.
      parameters:
        - description: File ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Delete a file
      tags:
        - files
    get:
      description: Returns file metadata if the caller is owner or admin.
      parameters:
        - description: File ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.FileResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get file metadata
      tags:
        - files
  /api/v1/files/{id}/download-signed:
    get:
      description: Returns a temporary signed URL for downloading a file.
      parameters:
        - description: File ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/presenter.SignedDownloadResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get signed download URL
      tags:
        - files
  /api/v1/files/public/{id}/download:
    get:
      description: Redirects to the public file URL if the file visibility is public.
      parameters:
        - description: File ID
          in: path
          name: id
          required: true
          type: string
      responses:
        "302":
          description: Found
        "403":
          description: Forbidden
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Public file download
      tags:
        - files
  /api/v1/files/upload:
    post:
      consumes:
        - multipart/form-data
      description: Uploads a file to object storage and stores metadata in database.
      parameters:
        - description: File binary
          in: formData
          name: file
          required: true
          type: file
        - description: Optional storage path prefix inside the configured bucket base path
          in: formData
          name: path
          type: string
        - description: private or public
          enum:
            - private
            - public
          in: formData
          name: visibility
          type: string
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/handlers.UploadFileResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Upload a file
      tags:
        - files
...
securityDefinitions:
  BearerAuth:
    description: Use `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"